new data collector mechanism
Now it is possible to select the your API JSON key to access your Google Sheets API. The special SF_PARSING folder creates in your Google Drive storage that shares the created files with your account (add your google email in order to make possible to create the shared folder).
Also the data checker is added - now the script checks if the file is existed or not and only after check - add new data to the file.